// packages/openverse-api-client/openverse_api_client-0.0.1.tar.gz/openverse_api_client-0.0.1/packages/openverse-api-client/src/types.ts
export interface _AudioByQuerySearchParams {
    category?: string | Array<string>
    length?: string | Array<string>
    peaks?: boolean
    page?: number
    page_size?: number
    source?: string | Array<string>
    excluded_source?: string | Array<string>
    license?: string | Array<string>
    license_type?: string | Array<string>
    tags?: string | Array<string>
    filter_dead?: boolean
    extension?: string | Array<string>
    include_sensitive_results?: boolean
    q: string
}

export interface SourceStats {
    source_name: string
    display_name: string
    source_url: string
    media_count: number
}

export interface _ImageByFieldSearchParams {
    category?: string | Array<string>
    aspect_ratio?: string | Array<string>
    size?: Array<string>
    page?: number
    page_size?: number
    source?: string | Array<string>
    excluded_source?: string | Array<string>
    license?: string | Array<string>
    license_type?: string | Array<string>
    tags?: string | Array<string>
    filter_dead?: boolean
    extension?: string | Array<string>
    include_sensitive_results?: boolean
    creator: string
    title: string
}

export interface ImageDetail {
    id: string
    title: string | null
    indexed_on: string
    foreign_landing_url: string | null
    url: string | null
    creator: string | null
    creator_url: string | null
    license: string
    license_version: string | null
    license_url: string | null
    provider: string | null
    source: string | null
    category: string | null
    filesize: string | null
    filetype: string | null
    tags: Array<Tag> | null
    attribution: string
    fields_matched: Array<string> | null
    mature: boolean
    thumbnail: string
    detail_url: string
    related_url: string
    height: number | null
    width: number | null
}

export interface _AudioByFieldSearchParams {
    category?: string | Array<string>
    length?: string | Array<string>
    peaks?: boolean
    page?: number
    page_size?: number
    source?: string | Array<string>
    excluded_source?: string | Array<string>
    license?: string | Array<string>
    license_type?: string | Array<string>
    tags?: string | Array<string>
    filter_dead?: boolean
    extension?: string | Array<string>
    include_sensitive_results?: boolean
    creator: string
    title: string
}

export interface AudioSet {
    title: string | null
    foreign_landing_url: string | null
    creator: string | null
    creator_url: string | null
    url: string | null
    filesize: number | null
    filetype: number | null
}

export interface AudioDetail {
    id: string
    title: string | null
    indexed_on: string
    foreign_landing_url: string | null
    url: string | null
    creator: string | null
    creator_url: string | null
    license: string
    license_version: string | null
    license_url: string | null
    provider: string | null
    source: string | null
    category: string | null
    filesize: string | null
    filetype: string | null
    tags: Array<Tag> | null
    attribution: string
    fields_matched: Array<string> | null
    mature: boolean
    thumbnail: string
    detail_url: string
    related_url: string
    audio_set: AudioSet | null
    duration: number | null
    bit_rate: number | null
    sample_rate: number | null
    waveform: string
    peaks: Array<number> | null
    alt_files: Record<string, any> | null
    genres: Array<string> | null
}

export interface Tag {
    name: string
    accuracy?: number | null
}

export interface _ImageByQuerySearchParams {
    category?: string | Array<string>
    aspect_ratio?: string | Array<string>
    size?: Array<string>
    page?: number
    page_size?: number
    source?: string | Array<string>
    excluded_source?: string | Array<string>
    license?: string | Array<string>
    license_type?: string | Array<string>
    tags?: string | Array<string>
    filter_dead?: boolean
    extension?: string | Array<string>
    include_sensitive_results?: boolean
    q: string
}


export interface RestAPI {
    "POST v1/auth_tokens/register/": {
        params: {
            name: string,
            description: string,
            email: string,
        },
    
        response: {
            name: string,
            client_id: string,
            client_secret: string,
        },
    }

    "POST v1/auth_tokens/token/": {
        params: {
            grant_type: 'client_credentials',
            client_id: string,
            client_secret: string,
        },
    
        response: {
            access_token: string,
            scope: string,
            expires_in: number,
            token_type: string,
        },
    }

    "GET v1/rate_limit/": {
        params: null,
        response: {
            requests_this_minute: number | null,
            requests_today: number | null,
            rate_limit_model: 'standard' | 'enhanced',
            verified: boolean,
        },
    }

    "GET v1/images/stats/": {
        params: null,
        response: Array<SourceStats>
    }

    "GET v1/audio/stats/": {
        params: null,
        response: Array<SourceStats>
    }

    "GET v1/images/": {
        params: _ImageByQuerySearchParams | _ImageByFieldSearchParams,
        response: {
            result_count: number,
            page_count: number,
            page_size: number,
            page: number,
            results: Array<ImageDetail>,
        },
    }

    "GET v1/audio/": {
        params: _AudioByQuerySearchParams | _AudioByFieldSearchParams,
        response: {
            result_count: number,
            page_count: number,
            page_size: number,
            page: number,
            results: Array<AudioDetail>,
        },
    }

    "GET v1/images/:identifier/related": {
        params: {
            identifier: string,
        },
    
        response: {
            result_count: number,
            page_count: number,
            page_size: number,
            page: number,
            results: Array<ImageDetail>,
        },
    }

    "GET v1/audio/:identifier/related": {
        params: {
            identifier: string,
        },
    
        response: {
            result_count: number,
            page_count: number,
            page_size: number,
            page: number,
            results: Array<AudioDetail>,
        },
    }

    "GET v1/images/:identifier/thumb/": {
        params: {
            identifier: string,
            full_size?: boolean,
            compressed?: boolean,
        },
    
        response: ReadableStream
    }

    "GET v1/audio/:identifier/thumb/": {
        params: {
            identifier: string,
            full_size?: boolean,
            compressed?: boolean,
        },
    
        response: ReadableStream
    }

    "GET v1/audio/:identifier/waveform/": {
        params: {
            identifier: string,
        },
    
        response: {
            len: number,
            points: Array<number>,
        },
    }

    "GET v1/images/oembed": {
        params: {
            url: string,
        },
    
        response: {
            version: string,
            type: string,
            width: number,
            height: number,
            title: string | null,
            author_name: string | null,
            author_url: string | null,
            license_url: string,
        },
    }
}

export const RestAPIMeta = Object.freeze({
    "POST v1/auth_tokens/register/": {
        contentType: "application/json",
        pathParams: [],
        jsonResponse: true,
    },

    "POST v1/auth_tokens/token/": {
        contentType: "application/x-www-form-urlencoded",
        pathParams: [],
        jsonResponse: true,
    },

    "GET v1/rate_limit/": {
        contentType: "application/json",
        pathParams: [],
        jsonResponse: true,
    },

    "GET v1/images/stats/": {
        contentType: "application/json",
        pathParams: [],
        jsonResponse: false,
    },

    "GET v1/audio/stats/": {
        contentType: "application/json",
        pathParams: [],
        jsonResponse: false,
    },

    "GET v1/images/": {
        contentType: "application/json",
        pathParams: [],
        jsonResponse: true,
    },

    "GET v1/audio/": {
        contentType: "application/json",
        pathParams: [],
        jsonResponse: true,
    },

    "GET v1/images/:identifier/related": {
        contentType: "application/json",
        pathParams: ["identifier"],
        jsonResponse: true,
    },

    "GET v1/audio/:identifier/related": {
        contentType: "application/json",
        pathParams: ["identifier"],
        jsonResponse: true,
    },

    "GET v1/images/:identifier/thumb/": {
        contentType: "application/json",
        pathParams: ["identifier"],
        jsonResponse: false,
    },

    "GET v1/audio/:identifier/thumb/": {
        contentType: "application/json",
        pathParams: ["identifier"],
        jsonResponse: false,
    },

    "GET v1/audio/:identifier/waveform/": {
        contentType: "application/json",
        pathParams: ["identifier"],
        jsonResponse: true,
    },

    "GET v1/images/oembed": {
        contentType: "application/json",
        pathParams: [],
        jsonResponse: true,
    },
} as const)